uC5282/clock/clock.c: Correct bsp_idle_thread() prototype

+++ bsps/m68k/uC5282/clock/clock.c
@@ -134,7 +134,7 @@ static void uC5282_tc_tick(void)
 /*
  * Provide our own version of the idle task
  */
-void * bsp_idle_thread(uint32_t ignored)
+void * bsp_idle_thread(uintptr_t ignored)
 {
   /* Atomic increment */
   for(;;)
